Make the remote connection a target of itself
